refactor: stores 폴더 barrel export 제거
