--- frontend/techpick-extension/src/stores/index.ts
@@ -1,3 +1,0 @@
-export { useTagStore } from './tagStore';
-export { useDeleteTagDialogStore } from './deleteTagDialogStore';
-export { useThemeStore } from './themeStore';
