+++ frontend/techpick-extension/src/types/CreatePickToUnclassifiedFolderRequestType.ts
@@ -1,5 +1,5 @@
-import { components } from '@/schema';
-import { ConcreteType } from './ConcreteType';
+import type { components } from '@/schema';
+import type { ConcreteType } from './ConcreteType';
 
 export type CreatePickToUnclassifiedFolderRequestType = ConcreteType<
   components['schemas']['baguni.api.application.pick.dto.PickApiRequest$CreateFromExtension']
